requirements.txt
aiogram==3.4.1
python-dotenv==1.0.1
gspread==6.2.1
google-auth==2.40.3
